style: collaborator 패키지로 관련 기능 이동 (#76)
