+++ src/main/java/com/listywave/collaborator/application/dto/CollaboratorSearchResponse.java
@@ -1,4 +1,4 @@
-package com.listywave.user.application.dto;
+package com.listywave.collaborator.application.dto;
 
 import java.util.List;
 import lombok.Builder;
