+++ src/main/java/com/listywave/user/presentation/UserController.java
@@ -5,14 +5,12 @@ import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import com.listywave.list.application.domain.CategoryType;
 import com.listywave.user.application.dto.AllUserListsResponse;
 import com.listywave.user.application.dto.AllUserResponse;
-import com.listywave.user.application.dto.CollaboratorSearchResponse;
 import com.listywave.user.application.dto.FollowingsResponse;
 import com.listywave.user.application.dto.RecommendUsersResponse;
 import com.listywave.user.application.dto.UserInfoResponse;
 import com.listywave.user.application.service.UserService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -83,15 +81,5 @@ public class UserController {
     ResponseEntity<List<RecommendUsersResponse>> getRecommendUsers() {
         List<RecommendUsersResponse> recommendUsers = userService.getRecommendUsers();
         return ResponseEntity.ok(recommendUsers);
-    }
-
-    @GetMapping("/collaborators")
-    ResponseEntity<CollaboratorSearchResponse> getCollaborators(
-            @RequestHeader(value = AUTHORIZATION, defaultValue = "") String accessToken,
-            @RequestParam(name = "search", defaultValue = "") String search,
-            Pageable pageable
-    ) {
-        CollaboratorSearchResponse collaborators = userService.getCollaborators(accessToken, search, pageable);
-        return ResponseEntity.ok(collaborators);
     }
 }
