+++ src/main/java/com/listywave/user/repository/user/custom/CustomUserRepository.java
@@ -1,9 +1,9 @@
 package com.listywave.user.repository.user.custom;
 
+import com.listywave.collaborator.application.dto.CollaboratorResponse;
 import com.listywave.list.application.domain.CategoryType;
 import com.listywave.list.application.domain.Lists;
 import com.listywave.user.application.domain.User;
-import com.listywave.user.application.dto.CollaboratorResponse;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
